\name{list_providers}
\alias{list_providers}
\title{Get location coordinates.}
\description{
\code{list_providers} returns some place's latitude and longitude. 
}
\value{
This function returns a list of possible providers to use in \code{\link{create_map}}.
}

\usage{
list_providers()
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca.
}
\examples{
list_providers()
}
