\name{locations}
\alias{locations}
\docType{data}
\title{
Data: Birthplaces locations of classical sociologists.
}
\description{
Matrix with birthplace locations of some sociologists.
}
\usage{data("locations")}
\format{
  A matrix with birthplace locations of 16 sociologists:
  \describe{
    \item{\code{lat}}{: latitude.}
    \item{\code{lon}}{: longitude.}
  }
}
\source{
Own elaboration from manuals of sociology and OpenStreetMap
}
\references{
See \code{\link{sociologists}.}
}
\examples{
data(locations)
}
\keyword{datasets}
