\name{get_location}
\alias{get_location}
\title{Get location coordinates.}
\description{
\code{get_location} returns some place's latitude and longitude. 
}
\value{
Latitude and longitude coordinates.
}

\usage{
get_location(x)
}
\arguments{
\item{x}{Some place string to search in OpenStreetMap.}
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca.
}
\examples{
get_location("Salamanca, Spain")
}
