\name{list_providers}
\alias{list_providers}
\title{Get location coordinates.}
\description{
List of all providers with their map variations. 
}
\value{
This function returns a list of available map providers for the \code{\link{create_map}} function.
}

\usage{
list_providers()
}
\author{
Modesto Escobar \email{modesto@usal.es}, Department of Sociology and Communication, University of Salamanca, Carlos Prieto \email{cprietos@usal.es}, and  David Barrios \email{metal@usal.es}, Bioinformatics Service, University of Salamanca.
}
\examples{
list_providers()
}
