\name{list_providers}
\alias{list_providers}
\title{Get location coordinates.}
\description{
\code{list_providers} returns some place's latitude and longitude. 
}
\value{
This function returns a list of available map providers for the \code{\link{create_map}} function.
}

\usage{
list_providers()
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca.
}
\examples{
list_providers()
}
